(GH-3) Declare assembly CLSCompliant. Change casing of namespaces from Bbt to BBT.
