+++ src/BBT.Maybe.Tests/TestData/ReferencedClass.cs
@@ -1,13 +1,13 @@
-﻿namespace Bbt.Maybe.Tests.TestData
+﻿namespace BBT.Maybe.Tests.TestData
 {
     /// <summary>
     /// Used for test purposes.
     /// </summary>
-    public struct ReferencedStruct
+    public class ReferencedClass
     {
         /// <summary>
         /// Gets or sets the owner of the 1-n relationship.
         /// </summary>
-        public BaseStruct ReferencingStruct { get; set; }
+        public BaseClass ReferencingClass { get; set; }
     }
 }
