+++ src/BBT.Maybe.Tests/TestData/ReferencedStruct.cs
@@ -1,13 +1,13 @@
-﻿namespace Bbt.Maybe.Tests.TestData
+﻿namespace BBT.Maybe.Tests.TestData
 {
     /// <summary>
     /// Used for test purposes.
     /// </summary>
-    public class ReferencedClass
+    public struct ReferencedStruct
     {
         /// <summary>
         /// Gets or sets the owner of the 1-n relationship.
         /// </summary>
-        public BaseClass ReferencingClass { get; set; }
+        public BaseStruct ReferencingStruct { get; set; }
     }
 }
