+++ src/BBT.Maybe.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.271
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bbt.Maybe", "Bbt.Maybe\Bbt.Maybe.csproj", "{2F38B821-95F8-4560-B08E-9796C44D0C86}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BBT.Maybe", "Bbt.Maybe\BBT.Maybe.csproj", "{2F38B821-95F8-4560-B08E-9796C44D0C86}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bbt.Maybe.Tests", "Bbt.Maybe.Tests\Bbt.Maybe.Tests.csproj", "{86D8402D-88CF-4887-A9BC-5662FC23BAFE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BBT.Maybe.Tests", "Bbt.Maybe.Tests\BBT.Maybe.Tests.csproj", "{86D8402D-88CF-4887-A9BC-5662FC23BAFE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
