+++ src/BBT.Maybe/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+﻿[assembly: System.CLSCompliant(true)]
