+++ src/BBT.Maybe/MaybeFactory.cs
@@ -1,4 +1,4 @@
-﻿namespace Bbt.Maybe
+﻿namespace BBT.Maybe
 {
     using System.Diagnostics.CodeAnalysis;
