+++ src/BBT.Maybe/NoneCase.cs
@@ -1,4 +1,4 @@
-﻿namespace Bbt.Maybe
+﻿namespace BBT.Maybe
 {
     using System;
